Cleaned up MemberHandshake logic.
The memberhandshake is a form of server control; messages that are only send
between servers and end up at the TcpServer logic.

This pr generalizes that concept and renames it to server control. This way we
can easily hook in new stuff like ping messages.
